split up match and user routes for clarity
