+++ srcs/backend/database/matches.sql
@@ -1,0 +1,14 @@
+CREATE TABLE matches (
+    id INTEGER PRIMARY KEY AUTOINCREMENT,
+    uuid TEXT NOT NULL UNIQUE,
+    player1_id INTEGER NOT NULL,
+    player2_id INTEGER NOT NULL,
+    status TEXT NOT NULL CHECK (status IN ('completed', 'interrupted')),
+    winner_id INTEGER,
+    start_time TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
+    end_time TIMESTAMP,
+    duration INTEGER,
+    FOREIGN KEY (player1_id) REFERENCES users(id),
+    FOREIGN KEY (player2_id) REFERENCES users(id),
+    FOREIGN KEY (winner_id) REFERENCES users(id)
+);
